+++ tests/test_masks.py
@@ -1,10 +1,6 @@
-import re
-
 import pytest
 
 from src.masks import get_mask_account, get_mask_card_number
-from src.processing import filter_by_state, sort_by_date
-from src.widget import get_date, mask_account_card
 
 
 @pytest.mark.parametrize("numbers", ["123456789012345", "12345678901234"])
@@ -53,48 +49,3 @@ def test_get_mask_account(number, expected_exception):
             get_mask_account(number)
     else:
         assert get_mask_account(number) == "**4305"
-
-
-@pytest.mark.parametrize(
-    "input_data, expected_output",
-    [
-        # Тесты для номеров карт
-        ("Visa Platinum 7000792289606361", "7000792289606361"),
-        ("MasterCard 1234567812345678", "1234567812345678"),
-        ("МИР 1234123412341234", "1234123412341234"),
-        ("Card 1234567890123456", "1234567890123456"),
-        # Тесты для номеров счетов
-        ("Счет 64686473678894779589", "64686473678894779589"),
-        ("Счет 12345678901234567890", "12345678901234567890"),
-        ("Счет 12345678901234567890", "12345678901234567890"),
-        # Граничные случаи и некорректные данные
-        ("", ""),
-    ],
-)
-def test_mask_account_card(input_data, expected_output):
-    """Параметризованный тест для функции mask_account_card"""
-    assert mask_account_card(input_data) == expected_output
-
-
-@pytest.mark.parametrize("date_info, expected_date", [("2025-02-11T02:26:18.671407", "11.02.2025")])
-def test_check_date_reformat(date_info, expected_date):
-    assert get_date(date_info) == expected_date
-
-
-@pytest.mark.parametrize("date_info, expected_exception", [("43563456346", ValueError)])
-def test_date_format_validity(date_info, expected_exception):
-    if not re.match(r"^\d{4}-\d{2}-\d{2}T", date_info) and expected_exception:
-        with pytest.raises(ValueError, match="Неверный формат даты"):
-            get_date(date_info)
-
-
-def test_sort_by_state(unsorted_list_of_transactions, executed_list):
-    assert filter_by_state(unsorted_list_of_transactions) == executed_list
-
-
-def test_sort_by_date(unsorted_list_of_transactions, sorted_by_day_list):
-    assert sorted_by_day_list, sort_by_date(unsorted_list_of_transactions) == sorted_by_day_list
-
-
-def test_sort_by_date_ascending(unsorted_list_of_transactions, sorted_by_day_list):
-    assert sort_by_date(unsorted_list_of_transactions, False) == sorted_by_day_list[::-1]
